+++ packages/shared/lib/core/wallet/utils/outputs/index.ts
@@ -1,5 +1,6 @@
 export * from './getAmountFromOutput'
 export * from './getExpirationDateFromOutput'
+export * from './getFoundryActivityData'
 export * from './getMetadataFromOutput'
 export * from './getNativeTokenFromOutput'
 export * from './getSenderFromOutput'
@@ -7,6 +8,7 @@ export * from './getRecipientAddressFromOutput'
 export * from './getRecipientFromOutput'
 export * from './getStorageDepositFromOutput'
 export * from './getTagFromOutput'
+export * from './getTransactionActivityData'
 export * from './isOutputAsync'
 export * from './preprocessGroupedOutputs'
 export * from './preprocessTransaction'
